+++ apps/web/src/api-clients/identity/user-profile/schema/constant.ts
@@ -2,3 +2,8 @@ export const MULTI_FACTOR_AUTH_TYPE = {
     OTP: 'OTP',
     EMAIL: 'EMAIL',
 } as const;
+
+export const MFA_STATE = {
+    ENABLED: 'ENABLED',
+    DISABLED: 'DISABLED',
+} as const;
