+++ apps/web/src/api-clients/identity/user/schema/api-verbs/create.ts
@@ -1,5 +1,5 @@
 import type { Tags } from '@/api-clients/_common/schema/model';
-import type { MultiFactorAuthType } from '@/api-clients/identity/user-profile/schema/type';
+import type { MfaState, MultiFactorAuthType } from '@/api-clients/identity/user-profile/schema/type';
 import type { AuthType } from '@/api-clients/identity/user/schema/type';
 
 
@@ -13,6 +13,6 @@ export interface UserCreateParameters {
     timezone?: string;
     tags?: Tags;
     reset_password?: boolean;
-    enforce_mfa?: boolean;
-    enforce_mfa_type?: MultiFactorAuthType; // only when enforce_mfa is true, this field is required
+    enforce_mfa_state?: MfaState;
+    enforce_mfa_type?: MultiFactorAuthType; // only when enforce_mfa_state is ENABLED, this field is required
 }
