+++ apps/web/src/api-clients/identity/workspace-user/schema/api-verbs/create.ts
@@ -1,4 +1,5 @@
 import type { Tags } from '@/api-clients/_common/schema/model';
+import type { MfaState, MultiFactorAuthType } from '@/api-clients/identity/user-profile/schema/type';
 import type { AuthType } from '@/api-clients/identity/user/schema/type';
 
 export interface WorkspaceUserCreateParameters {
@@ -11,5 +12,7 @@ export interface WorkspaceUserCreateParameters {
     timezone?: string;
     tags?: Tags;
     reset_password?: boolean;
+    enforce_mfa_state?: MfaState;
+    enforce_mfa_type?: MultiFactorAuthType;
     role_id: string;
 }
