+++ apps/web/src/services/iam/pages/UserMainPage.vue
@@ -12,7 +12,7 @@ import { useAuthorizationStore } from '@/store/authorization/authorization-store
 import { useGrantScopeGuard } from '@/common/composables/grant-scope-guard';
 import { usePageEditableStatus } from '@/common/composables/page-editable-status';
 
-import UserMFASettingModal from '@/services/iam/components/mfa/UserMFASettingModal.vue';
+import UserBulkMFASettingModal from '@/services/iam/components/mfa/UserBulkMFASettingModal.vue';
 import UserAssignToGroupModal from '@/services/iam/components/UserAssignToGroupModal.vue';
 import UserManagementAddModal from '@/services/iam/components/UserManagementAddModal.vue';
 import UserManagementFormModal from '@/services/iam/components/UserManagementFormModal.vue';
@@ -106,8 +106,8 @@ onUnmounted(() => {
         <user-assign-to-group-modal v-if="hasReadWriteAccess"
                                     @confirm="refreshUserList"
         />
-        <user-m-f-a-setting-modal v-if="hasReadWriteAccess"
-                                  @confirm="refreshUserList"
+        <user-bulk-m-f-a-setting-modal v-if="hasReadWriteAccess"
+                                       @confirm="refreshUserList"
         />
         <!-- <user-m-f-a-secret-key-delete-modal v-if="hasReadWriteAccess"
                                             @confirm="refreshUserList"
